Move worker and controller in channel directory
