+++ platoon/tests/functional/test_worker.py
@@ -8,7 +8,7 @@ import theano
 import numpy as np
 
 sys.path.append(os.path.join(os.path.dirname(__file__), '..', '..', '..'))
-from platoon.worker import Worker
+from platoon import Worker
 
 
 class TestWorker(unittest.TestCase):
